# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Starshine Drive website and internal CRM API
servers:
  - url: /api
...
paths:
...
  /crm/email-senders:
    get:
      operationId: listEmailSenderAccounts
      tags: [crm]
      responses:
        "200":
          description: Sender account list
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/EmailSenderAccount" }
    post:
      operationId: createEmailSenderAccount
      tags: [crm]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/EmailSenderAccountInput" }
      responses:
        "201":
          description: Sender account created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EmailSenderAccount" }
  /crm/email-senders/{id}:
    patch:
      operationId: updateEmailSenderAccount
      tags: [crm]
      parameters: [{ $ref: "#/components/parameters/Id" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/EmailSenderAccountUpdate" }
      responses:
        "200":
          description: Sender account updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EmailSenderAccount" }
    delete:
      operationId: deleteEmailSenderAccount
      tags: [crm]
      parameters: [{ $ref: "#/components/parameters/Id" }]
      responses:
        "204":
          description: Sender account deleted
  /crm/email-senders/{id}/test:
    post:
      operationId: sendEmailSenderTest
      tags: [crm]
      parameters: [{ $ref: "#/components/parameters/Id" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TestEmailRequest" }
      responses:
        "200":
          description: Test email result
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TestEmailResult" }
...
  /crm/email-marketing/recipients-preview:
    get:
      operationId: previewEmailMarketingRecipients
      tags: [crm]
      parameters:
        - name: source
          in: query
          required: true
          schema: { type: string, enum: [customers, leads] }
        - name: cities
          in: query
          schema: { type: string }
          description: Comma-separated city names (customers only).
        - name: states
          in: query
          schema: { type: string }
          description: Comma-separated state names (customers only).
        - name: industries
          in: query
          schema: { type: string }
          description: Comma-separated industries (leads only).
      responses:
        "200":
          description: Recipient count and sample for the given filters
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EmailMarketingRecipientsPreview" }
  /crm/email-marketing/campaigns:
    get:
      operationId: listEmailCampaigns
      tags: [crm]
      responses:
        "200":
          description: Bulk email campaign history
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/EmailCampaign" }
    post:
      operationId: createEmailCampaign
      tags: [crm]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/EmailCampaignInput" }
      responses:
        "201":
          description: Campaign created and sending started
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EmailCampaign" }
  /crm/email-marketing/campaigns/{id}:
    get:
      operationId: getEmailCampaign
      tags: [crm]
      parameters: [{ $ref: "#/components/parameters/Id" }]
      responses:
        "200":
          description: Campaign detail with per-recipient send log
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EmailCampaignDetail" }
...
components:
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: integer, minimum: 1 }
...
  schemas:
...
    TestEmailRequest:
      type: object
      required: [to]
      properties:
        to: { type: string, format: email }
    TestEmailResult:
      type: object
      required: [success]
      properties:
        success: { type: boolean }
        message: { type: ["string", "null"] }
    EmailSenderAccount:
      type: object
      required: [id, label, smtpHost, smtpPort, smtpSecure, smtpUser, fromEmail, active, createdAt, updatedAt]
      properties:
        id: { type: integer }
        label: { type: string }
        smtpHost: { type: string }
        smtpPort: { type: integer }
        smtpSecure: { type: boolean }
        smtpUser: { type: string }
        fromEmail: { type: string }
        fromName: { type: ["string", "null"] }
        active: { type: boolean }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    EmailSenderAccountInput:
      type: object
      required: [label, smtpHost, smtpUser, smtpPassword, fromEmail]
      properties:
        label: { type: string, minLength: 1 }
        smtpHost: { type: string, minLength: 1 }
        smtpPort: { type: integer }
        smtpSecure: { type: boolean }
        smtpUser: { type: string, minLength: 1 }
        smtpPassword: { type: string, minLength: 1 }
        fromEmail: { type: string, format: email }
        fromName: { type: string }
        active: { type: boolean }
    EmailSenderAccountUpdate:
      type: object
      properties:
        label: { type: string, minLength: 1 }
        smtpHost: { type: string, minLength: 1 }
        smtpPort: { type: integer }
        smtpSecure: { type: boolean }
        smtpUser: { type: string, minLength: 1 }
        smtpPassword: { type: string }
        fromEmail: { type: string, format: email }
        fromName: { type: ["string", "null"] }
        active: { type: boolean }
...
    EmailMarketingRecipientSample:
      type: object
      required: [email, name]
      properties:
        email: { type: string }
        name: { type: ["string", "null"] }
    EmailMarketingRecipientsPreview:
      type: object
      required: [count, sample]
      properties:
        count: { type: integer }
        sample:
          type: array
          items: { $ref: "#/components/schemas/EmailMarketingRecipientSample" }
    EmailCampaignFilters:
      type: object
      properties:
        cities: { type: array, items: { type: string } }
        states: { type: array, items: { type: string } }
        industries: { type: array, items: { type: string } }
    EmailCampaign:
      type: object
      required: [id, subject, body, recipientSource, filters, totalRecipients, sentCount, failedCount, status, createdAt, updatedAt]
      properties:
        id: { type: integer }
        subject: { type: string }
        body: { type: string }
        recipientSource: { type: string, enum: [customers, leads] }
        filters: { $ref: "#/components/schemas/EmailCampaignFilters" }
        totalRecipients: { type: integer }
        sentCount: { type: integer }
        failedCount: { type: integer }
        status: { type: string, enum: [sending, completed, failed] }
        createdBy: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    EmailCampaignInput:
      type: object
      required: [subject, body, recipientSource]
      properties:
        subject: { type: string, minLength: 1 }
        body: { type: string, minLength: 1 }
        recipientSource: { type: string, enum: [customers, leads] }
        cities: { type: string, description: "Comma-separated city names (customers only)." }
        states: { type: string, description: "Comma-separated state names (customers only)." }
        industries: { type: string, description: "Comma-separated industries (leads only)." }
    EmailCampaignRecipient:
      type: object
      required: [id, campaignId, recipientEmail, status, createdAt]
      properties:
        id: { type: integer }
        campaignId: { type: integer }
        recipientEmail: { type: string }
        recipientName: { type: ["string", "null"] }
        senderAccountId: { type: ["integer", "null"] }
        status: { type: string, enum: [pending, sent, failed] }
        error: { type: ["string", "null"] }
        sentAt: { type: ["string", "null"], format: date-time }
        createdAt: { type: string, format: date-time }
    EmailCampaignDetail:
      allOf:
        - { $ref: "#/components/schemas/EmailCampaign" }
        - type: object
          required: [recipients]
          properties:
            recipients:
              type: array
              items: { $ref: "#/components/schemas/EmailCampaignRecipient" }
...