# lib/api-zod/src/generated/types/emailCampaignFilters.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Starshine Drive website and internal CRM API
 * OpenAPI spec version: 0.1.0
 */

export interface EmailCampaignFilters {
  cities?: string[];
  states?: string[];
  industries?: string[];
}