# lib/api-zod/src/generated/types/emailSenderAccount.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Starshine Drive website and internal CRM API
 * OpenAPI spec version: 0.1.0
 */

export interface EmailSenderAccount {
  id: number;
  label: string;
  smtpHost: string;
  smtpPort: number;
  smtpSecure: boolean;
  smtpUser: string;
  fromEmail: string;
  /** @nullable */
  fromName?: string | null;
  active: boolean;
  createdAt: Date;
  updatedAt: Date;
}